# ApiCodeGenPoc/Controllers/NewWeatherForecastController.cs
using Contract.Requests;
using Contract.Responses;
using Microsoft.AspNetCore.Mvc;

namespace ApiCodeGenPoc.Controllers
{
    [ApiController]
    [Route("[controller]")]
    public class NewWeatherForecastController : ControllerBase
    {
        private static readonly string[] Summaries = new[]
        {
            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
        };

        private readonly ILogger<WeatherForecastController> _logger;

        public NewWeatherForecastController(ILogger<WeatherForecastController> logger)
        {
            _logger = logger;
        }

        [HttpPost]
        [ProducesResponseType(typeof(ICollection<WeatherForecastResponse>), StatusCodes.Status200OK)]
        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
        [ProducesResponseType(StatusCodes.Status403Forbidden)]
        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
        public ActionResult<IEnumerable<WeatherForecastResponse>> GetWeatherNew([FromBody] WeatherForecastRequest weatherForecastRequest)
        {
            var data = Enumerable.Range(1, 5).Select(index => new WeatherForecastResponse
            {
                Date = DateTime.Now.AddDays(index),
                TemperatureC = Random.Shared.Next(-20, 55),
                Summary = Summaries[Random.Shared.Next(Summaries.Length)]
            })
            .ToArray();

            return Ok(data);
        }
    }
}